Confirming architectural directions
My gut says, it's not yet this, but it's headed that way as new files or abstractions appear not to have been used (though some of the fine details I'd agree with):
https://en.wikipedia.org/wiki/Anti-pattern#Big_ball_of_mud

More precisely, the remaining new code is at its core the ROS2 demos, so while the choice of a ROS2 action results in a smoother implementation than the original pub/sub, it's not built in a way where collaboration is practical, such as the breaking choice of requiring a fork to modify pos files (and there's been no recent activity). So rUNSWift could be better served by just following the tutorials and benefitting from longer term potential collaboration.

    https://github.com/ros2/demos/tree/humble/action_tutorials
    https://github.com/ros2/demos/blob/rolling/action_tutorials/action_tutorials_cpp/README.md
    https://docs.ros2.org/latest/api/rclpy/api/actions.html
    https://docs.ros.org/en/humble/Tutorials/Beginner-CLI-Tools/Understanding-ROS2-Actions/Understanding-ROS2-Actions.html
    https://docs.ros.org/en/humble/Tutorials/Intermediate/Creating-an-Action.html
    https://docs.ros.org/en/humble/Tutorials/Intermediate/Writing-an-Action-Server-Client/Cpp.html
    https://docs.ros.org/en/humble/Tutorials/Intermediate/Writing-an-Action-Server-Client/Py.html